commiting work so far on table and database creation/server.js
