--- .env.EXAMPLE
@@ -1,3 +1,0 @@
-DB_NAME=tech_blog_db
-DB_USER=
-DB_PW=
