+++ controllers/api/index.js
@@ -1,11 +1,9 @@
 const router = require('express').Router();
 
 const userRoutes = require('./user-routes.js');
-const postRoutes = require('./post-routes');
-const commentRoutes = require('./comment-routes');
+// const jokeRoutes = require('./joke-routes');
 
 router.use('/user', userRoutes);
-router.use('/post', postRoutes);
-router.use('/comment', commentRoutes);
+// router.use('/post', jokeRoutes);
 
 module.exports = router;
