+++ controllers/api/joke-routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Post } = require('../../models/');
+const { newPost } = require('../../models');
 const withAuth = require('../../utils/auth');
 
 router.post('/', withAuth, async (req, res) => {
