+++ controllers/home-routes.js
@@ -1,68 +1,38 @@
 const router = require('express').Router();
-const { Post, Comment, User } = require('../models/');
+const { User } = require('../models/');
 
 // get all posts for homepage
-router.get('/', async (req, res) => {
-  try {
-    // we need to get all Posts and include the User for each (change lines 8 and 9)
-    const postData = await SomeModel.someSequelizeMethod({
-      include: [SomeOtherModel],
+router.get('/', (req, res) => {
+  console.log('======================');
+  Post.findAll({
+    attributes: [
+      'id',
+      'post_url',
+      'title',
+      'created_at',
+    ],
+    // include: [
+    //   {
+    //     model: Comment,
+    //     attributes: ['id', 'comment_text', 'post_id', 'user_id', 'created_at'],
+    //     include: {
+    //       model: User,
+    //       attributes: ['username']
+    //     }
+    //   },
+    //   {
+    //     model: User,
+    //     attributes: ['username']
+    //   }
+    // ]
+  })
+    .then(dbPostData => {
+      const posts = dbPostData.map(post => post.get({ plain: true }));
+
+      res.render('homepage', { posts });
+    })
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
     });
-    // serialize the data
-    const posts = postData.map((post) => post.get({ plain: true }));
-    // we should render all the posts here
-    res.render('hmmmm what view should we render?', { posts });
-  } catch (err) {
-    res.status(500).json(err);
-  }
 });
-
-// get single post
-router.get('/post/:id', async (req, res) => {
-  try {
-    // what should we pass here? we need to get some data passed via the request body (something.something.id?)
-    // change the model below, but not the findByPk method.
-    const postData = await SomeModel.findByPk(????, {
-      // helping you out with the include here, no changes necessary
-      include: [
-        User,
-        {
-          model: Comment,
-          include: [User],
-        },
-      ],
-    });
-
-    if (postData) {
-      // serialize the data
-      const post = postData.get({ plain: true });
-      // which view should we render for a single-post?
-      res.render('hmmmm what view should we render?', { post });
-    } else {
-      res.status(404).end();
-    }
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-// giving you the login and signup route pieces below, no changes needed.
-router.get('/login', (req, res) => {
-  if (req.session.loggedIn) {
-    res.redirect('/');
-    return;
-  }
-
-  res.render('login');
-});
-
-router.get('/signup', (req, res) => {
-  if (req.session.loggedIn) {
-    res.redirect('/');
-    return;
-  }
-
-  res.render('signup');
-});
-
-module.exports = router;
