+++ controllers/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
-const apiRoutes = require('./api/');
-const homeRoutes = require('./home-routes.js');
+// const apiRoutes = require('./api');
+// const homeRoutes = require('./home-routes.js');
 const dashboardRoutes = require('./dashboard-routes.js');
 
 router.use('/', homeRoutes);
