+++ db/schema.sql
@@ -1,3 +1,3 @@
-DROP DATABASE IF EXISTS tech_blog_db;
+DROP DATABASE IF EXISTS jokr_db;
 
-CREATE DATABASE tech_blog_db;
+CREATE DATABASE jokr_db;
