+++ models/index.js
@@ -1,24 +1,17 @@
-const User = require('./User');
-const Post = require('./Post');
-const Comment = require('./Comment');
+const User = require("./User");
+// const Joke = require("./Joke");
 
-Post.belongsTo(User, {
-  foreignKey: 'userId',
-  onDelete: 'CASCADE'
-});
+// Joke.belongsTo(User, {
+//   foreignKey: "user_id",
+//   onDelete: "CASCADE",
+// });
 
-Post.hasMany(Comment, {
-  foreignKey: 'postId',
-  onDelete: 'CASCADE'
-});
-
-Comment.belongsTo(User, {
-  foreignKey: 'userId',
-  onDelete: 'CASCADE'
-});
+// User.hasMany(Joke, {
+//   foreignKey: "user_id",
+//   onDelete: "CASCADE",
+// });
 
 module.exports = {
-  User,
-  Comment,
-  Post
+  User
+  // Joke
 };
